pyproject.toml
[project]
name = "nsct-toolbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pillow>=11.3.0",
]
